+++ hi-ml-azure/run_requirements.txt
@@ -1,4 +1,5 @@
 azure-ai-ml>=1.1.1
+azure-cli>=2.30.0
 azureml-core>=1.42.0
 azureml-dataset-runtime[fuse]>=1.42.0
 azureml-mlflow>=1.42.0
