+++ hi-ml-azure/src/health_azure/utils.py
@@ -39,7 +39,6 @@ import param
 
 from azureml._restclient.constants import RunStatus
 from azureml.core import Environment, Experiment, Run, Workspace, get_run
-from azureml.core.authentication import InteractiveLoginAuthentication, ServicePrincipalAuthentication
 from azureml.core.conda_dependencies import CondaDependencies
 from azureml.core.run import _OfflineRun
 from azureml.data.azure_storage_datastore import AzureBlobDatastore
@@ -48,17 +47,10 @@ from azure.ai.ml import MLClient
 from azure.ai.ml.entities import Job
 from azure.ai.ml.entities import Workspace as WorkspaceV2
 from azure.ai.ml.entities import Environment as EnvironmentV2
-from azure.core.credentials import TokenCredential
-from azure.core.exceptions import ClientAuthenticationError, ResourceNotFoundError
-from azure.identity import (
-    ClientSecretCredential,
-    DeviceCodeCredential,
-    DefaultAzureCredential,
-    InteractiveBrowserCredential,
-)
+from azure.core.exceptions import ResourceNotFoundError
 
 from health_azure.argparsing import EXPERIMENT_RUN_SEPARATOR, RunIdOrListParam, determine_run_id_type
-
+from health_azure.auth import get_authentication, get_credential, get_secret_from_environment
 
 logger = logging.getLogger(__name__)
 
@@ -69,11 +61,6 @@ DEFAULT_UPLOAD_TIMEOUT_SECONDS: int = 36_000  # 10 Hours
 # The version to use when creating an AzureML Python environment. We create all environments with a unique hashed
 # name, hence version will always be fixed
 ENVIRONMENT_VERSION = "1"
-
-# Environment variables used for authentication
-ENV_SERVICE_PRINCIPAL_ID = "HIML_SERVICE_PRINCIPAL_ID"
-ENV_SERVICE_PRINCIPAL_PASSWORD = "HIML_SERVICE_PRINCIPAL_PASSWORD"
-ENV_TENANT_ID = "HIML_TENANT_ID"
 
 # Environment variables used for workspace selection
 ENV_RESOURCE_GROUP = "HIML_RESOURCE_GROUP"
@@ -441,51 +428,6 @@ def fetch_run_for_experiment(experiment_to_recover: Experiment, run_id: str) -> 
         raise Exception(
             f"Run {run_id} not found for experiment: {experiment_to_recover.name}. Available runs are: {available_ids}"
         )
-
-
-def get_authentication() -> Union[InteractiveLoginAuthentication, ServicePrincipalAuthentication]:
-    """
-    Creates a service principal authentication object with the application ID stored in the present object. The
-    application key is read from the environment.
-
-    :return: A ServicePrincipalAuthentication object that has the application ID and key or None if the key is not
-        present
-    """
-    service_principal_id = get_secret_from_environment(ENV_SERVICE_PRINCIPAL_ID, allow_missing=True)
-    tenant_id = get_secret_from_environment(ENV_TENANT_ID, allow_missing=True)
-    service_principal_password = get_secret_from_environment(ENV_SERVICE_PRINCIPAL_PASSWORD, allow_missing=True)
-    # Check if all 3 environment variables are set
-    if service_principal_id and tenant_id and service_principal_password:
-        logger.info(
-            "Found environment variables for Service Principal authentication: First characters of App ID "
-            f"are {service_principal_id[:8]}... in tenant {tenant_id[:8]}..."
-        )
-        return ServicePrincipalAuthentication(
-            tenant_id=tenant_id,
-            service_principal_id=service_principal_id,
-            service_principal_password=service_principal_password,
-        )
-    logger.info(
-        "Using interactive login to Azure. To use Service Principal authentication, set the environment "
-        f"variables {ENV_SERVICE_PRINCIPAL_ID}, {ENV_SERVICE_PRINCIPAL_PASSWORD}, and {ENV_TENANT_ID}"
-    )
-    return InteractiveLoginAuthentication()
-
-
-def get_secret_from_environment(name: str, allow_missing: bool = False) -> Optional[str]:
-    """
-    Gets a password or key from the secrets file or environment variables.
-
-    :param name: The name of the environment variable to read. It will be converted to uppercase.
-    :param allow_missing: If true, the function returns None if there is no entry of the given name in any of the
-        places searched. If false, missing entries will raise a ValueError.
-    :return: Value of the secret. None, if there is no value and allow_missing is True.
-    """
-    name = name.upper()
-    value = os.environ.get(name, None)
-    if not value and not allow_missing:
-        raise ValueError(f"There is no value stored for the secret named '{name}'")
-    return value
 
 
 def to_azure_friendly_string(x: Optional[str]) -> Optional[str]:
@@ -1842,125 +1784,6 @@ def check_is_any_of(message: str, actual: Optional[str], valid: Iterable[Optiona
     if actual not in valid:
         all_valid = ", ".join(["<None>" if v is None else v for v in valid])
         raise ValueError("{} must be one of [{}], but got: {}".format(message, all_valid, actual))
-
-
-def _validate_credential(credential: TokenCredential) -> None:
-    """
-    Validate credential by attempting to get token. If authentication has been successful, get_token
-    will succeed. Otherwise an exception will be raised
-
-    :param credential: The credential object to validate.
-    """
-    credential.get_token("https://management.azure.com/.default")
-
-
-def _get_legitimate_service_principal_credential(
-    tenant_id: str, service_principal_id: str, service_principal_password: str
-) -> TokenCredential:
-    """
-    Create a ClientSecretCredential given a tenant id, service principal id and password
-
-    :param tenant_id: The Azure tenant id.
-    :param service_principal_id: The id of an existing Service Principal.
-    :param service_principal_password: The password of an existing Service Principal.
-    :raises ValueError: If the credential cannot be validated (i.e. authentication was unsucessful).
-    :return: The validated credential.
-    """
-    cred = ClientSecretCredential(
-        tenant_id=tenant_id, client_id=service_principal_id, client_secret=service_principal_password
-    )
-    try:
-        _validate_credential(cred)
-        return cred
-    except ClientAuthenticationError as e:
-        raise ValueError(
-            f"Found environment variables for {ENV_SERVICE_PRINCIPAL_ID}, "
-            f"{ENV_SERVICE_PRINCIPAL_PASSWORD}, and {ENV_TENANT_ID} but was "
-            f"not able to authenticate: {e}"
-        )
-
-
-def _get_legitimate_device_code_credential() -> Optional[TokenCredential]:
-    """
-    Create a DeviceCodeCredential for interacting with Azure resources. If the credential can't be
-    validated, return None.
-
-    :return: A valid Azure credential.
-    """
-    cred = DeviceCodeCredential(timeout=60)
-    try:
-        _validate_credential(cred)
-        return cred
-    except ClientAuthenticationError:
-        return None
-
-
-def _get_legitimate_default_credential() -> Optional[TokenCredential]:
-    """
-    Create a DefaultAzure credential for interacting with Azure resources and validates it.
-
-    :return: A valid Azure credential.
-    """
-    cred = DefaultAzureCredential(timeout=60)
-    _validate_credential(cred)
-    return cred
-
-
-def _get_legitimate_interactive_browser_credential() -> Optional[TokenCredential]:
-    """
-    Create an InteractiveBrowser credential for interacting with Azure resources. If the credential can't be
-    validated, return None.
-
-    :return: A valid Azure credential.
-    """
-    cred = InteractiveBrowserCredential(timeout=60)
-    try:
-        _validate_credential(cred)
-        return cred
-    except ClientAuthenticationError:
-        return None
-
-
-def get_credential() -> Optional[TokenCredential]:
-    """
-    Get a credential for authenticating with Azure. There are multiple ways to retrieve a credential.
-    If environment variables pertaining to details of a Service Principal are available, those will be used
-    to authenticate. If no environment variables exist, and the script is not currently
-    running inside of Azure ML or another Azure agent, will attempt to retrieve a credential via a
-    device code (which requires the user to visit a link and enter a provided code). If this fails, or if running in
-    Azure, DefaultAzureCredential will be used which iterates through a number of possible authentication methods
-    including identifying an Azure managed identity, cached credentials from VS code, Azure CLI, Powershell etc.
-    Otherwise returns None.
-
-    :return: Any of the aforementioned credentials if available, else None.
-    """
-    service_principal_id = get_secret_from_environment(ENV_SERVICE_PRINCIPAL_ID, allow_missing=True)
-    tenant_id = get_secret_from_environment(ENV_TENANT_ID, allow_missing=True)
-    service_principal_password = get_secret_from_environment(ENV_SERVICE_PRINCIPAL_PASSWORD, allow_missing=True)
-    if service_principal_id and tenant_id and service_principal_password:
-        logger.info(
-            "Found environment variables for Service Principal authentication: First characters of App ID "
-            f"are {service_principal_id[:8]}... in tenant {tenant_id[:8]}..."
-        )
-        return _get_legitimate_service_principal_credential(tenant_id, service_principal_id, service_principal_password)
-
-    try:
-        cred = _get_legitimate_default_credential()
-        if cred is not None:
-            return cred
-    except ClientAuthenticationError:
-        cred = _get_legitimate_device_code_credential()
-        if cred is not None:
-            return cred
-
-        cred = _get_legitimate_interactive_browser_credential()
-        if cred is not None:
-            return cred
-
-    raise ValueError(
-        "Unable to generate and validate a credential. Please see Azure ML documentation"
-        "for instructions on different options to get a credential"
-    )
 
 
 def get_ml_client(
